add owner to other routes

+++ controllers/contacts.js
@@ -1,9 +1,9 @@
 const Contact = require("../models/contact");
 const { HttpError, ctrlWrapper } = require("../helpers");
 
-const listContacts = async (req, res, next) => {
+const listContacts = async (req, res) => {
   const { _id: owner } = req.user;
-  const { page = 1, limit = 20, favorite = false } = req.query;
+  const { page = 1, limit = 20, favorite } = req.query;
   const skip = (page - 1) * limit;
   if (favorite === "true") {
     const result = await Contact.find({ owner, favorite: true }, "", {
@@ -19,46 +19,58 @@ const listContacts = async (req, res, next) => {
   res.json(result);
 };
 
-const getContactById = async (req, res, next) => {
+const getContactById = async (req, res) => {
+  const { _id: owner } = req.user;
   console.log(req.params);
   const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
+  const result = await Contact.findById({ _id: contactId, owner });
   if (!result) {
     throw HttpError(404, "Not Found");
   }
   res.json(result);
 };
 
-const addContact = async (req, res, next) => {
+const addContact = async (req, res) => {
   const { _id: owner } = req.user;
   const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
 
-const removeContact = async (req, res, next) => {
+const removeContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
+  const result = await Contact.findByIdAndRemove({ _id: contactId, owner });
   if (!result) {
     throw HttpError(404, "Not Found");
   }
   res.json({ message: "contact deleted" });
 };
 
-const updateContact = async (req, res, next) => {
+const updateContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
+  const result = await Contact.findByIdAndUpdate(
+    { _id: contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
   if (!result) {
     throw HttpError(404, "Not Found");
   }
   res.json(result);
 };
-const updateStatusContact = async (req, res, next) => {
+const updateStatusContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
+  const result = await Contact.findByIdAndUpdate(
+    { _id: contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
   if (!result) {
     throw HttpError(404, "Not Found");
   }
